panditji-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
email-validator
cloudinary
resend
sentry-sdk[fastapi]
slowapi
python-dotenv
razorpay
httpx
sqlalchemy[asyncio]
